pyproject.toml
[tool.poetry]
name = "ai-news-pipeline"
version = "0.1.0"
description = "AI-powered automated news processing pipeline"
authors = ["Jansaida Shaik <jansaidashaik1995@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
httpx = "^0.25.0"
feedparser = "^6.0.10"
openai = "^1.2.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
aiohttp = "^3.8.6"
asyncio = "^3.4.3"
pillow = "^11.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
black = "^23.10.0"
isort = "^5.12.0"
mypy = "^1.6.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"